refactor: extract shared payment functions to utils
- use cents-dollars converters on PaymentPageBlock + PaymentContent too
